Breakup tests into separate files
